+++ src/view/statistics.js
@@ -1,0 +1,7 @@
+const createStatisticsTemplate = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
+
+export {createStatisticsTemplate};
